imob_chatbot/requirements.txt
Flask
requests
google-auth
google-auth-oauthlib
twilio
python-dotenv
nohup
ngrok